Rename file as a place of arxml model elements
