+++ library/arxml_model.cpp
@@ -2,7 +2,7 @@
 // Created by Paweł Jarosz on 09.11.2023.
 //
 
-#include <arxml/foo.hpp>
+#include <arxml/arxml_model.hpp>
 
 int foo() {
     return 2;
